src/requirements.txt
# AMS - Agent Memory Scaffolding Requirements
# Python 3.10+ required

# Core Framework
dspy-ai>=2.5.0
pydantic>=2.0.0

# LLM Providers
openai>=1.0.0
litellm>=1.0.0

# Embeddings & Retrieval
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Data Processing
python-dateutil>=2.8.0

# Optional: For evaluation
rouge-score>=0.1.2
nltk>=3.8.0
bert-score>=0.3.13

